requirements.txt
aiogram>=3.0
pydantic-settings
aiosqlite
python-dotenv
python-dateutil
